requirements.txt
google-generativeai
streamlit
pandas
elevenlabs